+++ base/Member.js
@@ -19,8 +19,7 @@ module.exports = mongoose.model("Member", new mongoose.Schema({
     /* COOLDOWN */
     cooldowns: { type: Object, default: {
         work: 0,
-        rep: 0,
-        steal: 0
+        rob: 0
     }},
 
     /* OTHER INFORMATIONS */