+++ docs/changelog.md
@@ -7,6 +7,24 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 *   `serverinfo` command now displays the number of boosts of the server
+*   New `achievements` command
+
+## [4.5.10]
+
+### Added
+
+*   New `mention-role` command
+
+### Fixed
+
+*   If a member is muted and leaves and returns, he will always be muted
+*   User game is now displayed in the userinfo command
+*   Fix `invitations` command
+*   Rep cooldown is now user based and not member based
+*   `steal` command was renamed to `rob`
+*   `minimize` command was renamed to `shorturl`
+*   `ban` command now accepts ids
+*   Fixes some other minor bugs
 
 ## [4.5.0]